Refactor surrogate model evaluation CLI to use configuration parameters. Still need to finish the train_surrogate CLI refactor.
