--- bone_remodelling/surrogate_model/ensemble.py
@@ -1,124 +1,0 @@
-"""Loads an ensemble of surrogate models and returns mean and standard deviation predictions."""
-
-import logging
-from pathlib import Path
-
-import numpy as np
-import torch
-
-from bone_remodelling.surrogate_model.loader import SurrogateModelLoader
-from bone_remodelling.surrogate_model.neural_networks.neural_network import (
-    SurrogateModel,
-)
-from bone_remodelling.surrogate_model.normalizor import (
-    load_normalization_params,
-    normalize_data,
-    unnormalize_data,
-)
-
-logger = logging.getLogger(__name__)
-
-
-def load_ensemble_models(
-    model_paths: list[Path],
-    model_class: type[SurrogateModel],
-    model_loader: type[SurrogateModelLoader],
-) -> tuple[
-    list[SurrogateModel],
-    list[np.ndarray],
-    list[np.ndarray],
-    list[np.ndarray],
-    list[np.ndarray],
-]:
-    """Load multiple surrogate models from specified paths."""
-    models = []
-    x_means = []
-    x_stds = []
-    y_means = []
-    y_stds = []
-
-    for model_path in model_paths:
-        logger.info(f"Loading model from {model_path}")
-        model = model_loader(model_path, model_class).model
-        device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
-        model = model.to(device).eval()
-
-        x_mean, x_std, y_mean, y_std = load_normalization_params(
-            path=model_path.with_suffix(".npz"),
-        )
-
-        models.append(model)
-        x_means.append(x_mean)
-        x_stds.append(x_std)
-        y_means.append(y_mean)
-        y_stds.append(y_std)
-
-    return models, x_means, x_stds, y_means, y_stds
-
-
-def predict_with_ensemble(
-    models: list[SurrogateModel],
-    x: np.ndarray,
-    x_normalizations: tuple[list[np.ndarray], list[np.ndarray]],
-    y_normalizations: tuple[list[np.ndarray], list[np.ndarray]],
-) -> tuple[np.ndarray, np.ndarray]:
-    """Predict using an ensemble of models and return mean and standard deviation."""
-    predictions = []
-    x_means, x_stds = x_normalizations
-    y_means, y_stds = y_normalizations
-    for model, x_mean, x_std, y_mean, y_std in zip(
-        models,
-        x_means,
-        x_stds,
-        y_means,
-        y_stds,
-    ):
-        prediction = surrogate_model_forward(model, (x_mean, x_std), (y_mean, y_std), x)
-        predictions.append(prediction)
-
-    predictions = np.array(predictions)
-    mean_prediction = np.mean(predictions, axis=0)
-    std_prediction = np.std(predictions, axis=0)
-
-    return mean_prediction, std_prediction
-
-
-def surrogate_model_forward(
-    model: type[SurrogateModel],
-    x_normalization: tuple[np.ndarray, np.ndarray],
-    y_normalization: tuple[np.ndarray, np.ndarray],
-    force_profiles: np.ndarray,
-) -> np.ndarray:
-    """Forward pass through the surrogate model.
-
-    Args:
-    ----
-        model (SurrogateModel): The surrogate model to use for prediction.
-        x_normalization (tuple[np.ndarray, np.ndarray]): Mean and std for input normalization.
-        y_normalization (tuple[np.ndarray, np.ndarray]): Mean and std for output unnormalization.
-        force_profiles (np.ndarray): Input force profiles.
-
-    Returns:
-    -------
-        np.ndarray: The predicted density from the surrogate model.
-
-    """
-    x_mean, x_std = x_normalization
-    force_profile, _, _ = normalize_data(force_profiles, x_mean, x_std)
-
-    # forward pass through the surrogate model
-    device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
-    with torch.no_grad():
-        # Support batch predictions
-        force_profile_tensor = torch.from_numpy(
-            force_profile.reshape(force_profile.shape[0], -1).astype(np.float32)
-            if force_profile.ndim > 1
-            else force_profile.reshape(1, -1).astype(np.float32),
-        ).to(device)
-
-        density_tensor = model(force_profile_tensor)
-
-        y_mean, y_std = y_normalization
-        density_np = density_tensor.detach().cpu().numpy()
-        surrogate_density: np.ndarray = unnormalize_data(density_np, y_mean, y_std)
-    return surrogate_density
